requirements.txt
click==8.0.1
colorama==0.4.4
Flask==2.0.1
gunicorn==20.0.4
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Werkzeug==2.0.1
mysql-connector-python
flask-cors